+++ src/app/shared/services/counter-service/counter.service.ts
@@ -8,26 +8,38 @@ import {
 import { Counter } from "./../../model/counter";
 import { map } from "rxjs/operators";
 import { AuthService } from "../auth/auth.service";
+import { AngularFireAuth } from "@angular/fire/auth";
+import { Observable } from "rxjs";
 
 @Injectable()
 export class CounterService {
   collectionEndpoint: string;
-  counters: AngularFirestoreCollection<Counter>;
+  counterCollection: AngularFirestoreCollection<Counter>;
+  counters: Observable<any>;
 
-  constructor(private db: AngularFirestore, public authService: AuthService) {
-    this.collectionEndpoint = this.buildCollectionEndpoint();
-    this.counters = db.collection<Counter>(this.collectionEndpoint, ref =>
-      ref.orderBy("name")
-    );
+  constructor(
+    private angularFireAuth: AngularFireAuth,
+    private db: AngularFirestore
+  ) {
+    angularFireAuth.authState.subscribe(userData => {
+      if (userData) {
+        this.collectionEndpoint = this.buildCollectionEndpoint(userData);
+        this.counterCollection = db.collection<Counter>(
+          this.collectionEndpoint,
+          ref => ref.orderBy("name")
+        );
+        this.counters = this.getCounters();
+      }
+    });
   }
 
-  buildCollectionEndpoint() {
-    const pathStrings = ["users", this.authService.userData.uid, "counters"];
+  buildCollectionEndpoint(userData: firebase.User) {
+    const pathStrings = ["users", userData.uid, "counters"];
     return pathStrings.join("/");
   }
 
   getCounters() {
-    return this.counters.snapshotChanges().pipe(
+    return this.counterCollection.snapshotChanges().pipe(
       map(actions =>
         actions.map(a => {
           const data = a.payload.doc.data() as Counter;
@@ -41,11 +53,11 @@ export class CounterService {
   addCounter(name: string) {
     const id = this.db.createId();
     const counter = new Counter(id, name, 0, [], new Date());
-    this.counters.doc(id).set(Object.assign({}, counter));
+    this.counterCollection.doc(id).set(Object.assign({}, counter));
   }
 
   updateCounter(id: string, update: Partial<Counter>) {
-    this.counters.doc(id).update(update);
+    this.counterCollection.doc(id).update(update);
   }
 
   deleteCounter(counterId: string) {